requirements.txt
PyQt5==5.15.9
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0
